Add files via upload
Assignment 3:The purpose of this assignment is to use Naive Bayes for classification. 
This folder consists of the "assignment3.Rmd(R markdown)" file with the operations like Naive Bayes for classification performed on the  data set "UniversalBank.csv" file as per Assignment  instructions. The output is displayed in the files with different formats like HTML file "assignment3.html" and PDF file "assignment3.pdf"